API

API for Odrpack.

Odrpack.OdrResult — Type

OdrResult

Results of an Orthogonal Distance Regression (ODR) computation.

Fields

beta::Vector{Float64}: Estimated parameters of the model.

delta::Union{Vector{Float64},Matrix{Float64}}: Differences between the observed and fitted x values.

eps::Union{Vector{Float64},Matrix{Float64}}: Differences between the observed and fitted y values.

xplusd::Union{Vector{Float64},Matrix{Float64}}: Adjusted x values after fitting, x + delta.

yest::Union{Vector{Float64},Matrix{Float64}}: Estimated y values corresponding to the fitted model, y + eps.

sd_beta::Vector{Float64}: Standard deviations of the estimated parameters.

cov_beta::Matrix{Float64}: Covariance matrix of the estimated parameters.

res_var::Float64: Residual variance, indicating the variance of the residuals.

nfev::Int: Number of function evaluations during the fitting process.

njev::Int: Number of Jacobian evaluations during the fitting process.

niter::Int: Number of iterations performed in the optimization process.

irank::Int: Rank of the Jacobian matrix at the solution.

inv_condnum::Float64: Inverse of the condition number of the Jacobian matrix.

info::Int: Status code of the fitting process (e.g., success or failure).

stopreason::String: Message indicating the reason for stopping.

success::Bool: Whether the fitting process was successful.

sum_square::Float64: Sum of squared residuals (including both delta and eps).

sum_square_delta::Float64: Sum of squared differences between observed and fitted x values.

sum_square_eps::Float64: Sum of squared differences between observed and fitted y values.

_iwork::Vector{Int32}: Integer workspace array used internally by odrpack. Typically for advanced debugging.

_rwork::Vector{Float64}: Floating-point workspace array used internally by odrpack. Typically for advanced debugging.

Odrpack.OdrStop — Type

OdrStop <: Exception

Exception to stop the regression.

This exception can be raised in the model function or its Jacobians to stop the regression process.

Odrpack.Thunk — Type

Thunk

A mutable container used to pass around the model function, jacobians, and related information without creating a closure. cFunction does not support closures in certain platforms (e.g., macOS ARM).

Fields

f!::Function The main function to be evaluated.

jac_beta!::Union{Function,Nothing} Optional Jacobian with respect to parameter(s) β.

jac_x!::Union{Function,Nothing} Optional Jacobian with respect to input(s) x.

x_is_2d::Bool Whether the input x is treated as a matrix (true) or a vector (false).

y_is_2d::Bool Whether the output y = f(x, β) is treated as a matrix (true) or a vector (false).

Odrpack.odr_callback! — Method

Callback function invoked from odrpack to evaluate the model function and, optionally, its Jacobians. The actual julia functions are stored in the Thunk struct, which is passed as a void pointer. This is used to call the appropriate functions without creating closures.

Odrpack.odr_fit — Method

odr_fit(f!, xdata, ydata, beta0; kwargs...) -> OdrResult

Solve a weighted orthogonal distance regression (ODR) problem, also known as errors-in-variables regression.

Arguments

f!::Function: Function to be fitted, with the signature f!(x, beta, y). It must modify y in-place to have the same shape as ydata.

xdata::VecOrMat{Float64}: Array of shape (n,) or (n, m) containing the observed values of the explanatory variable(s).

ydata::VecOrMat{Float64}: Array of shape (n,) or (n, q) containing the observed values of the response variable(s). When the model is explicit, ydata must contain a value for each observation. To ignore specific values (e.g., missing data), set the corresponding entry in weight_y to zero. When the model is implicit, ydata is not used (but must be defined).

beta0::Vector{Float64}: Array of shape (np,) with the initial guesses of the model parameters, within the optional bounds specified by bounds.

Keywords

weight_x::Union{Float64,Vector{Float64},Matrix{Float64},Array{Float64,3},Nothing}=nothing: Scalar or array specifying how the errors on xdata are to be weighted. If weight_x is a scalar, then it is used for all data points. If weight_x is an array of shape (n,) and m==1, then weight_x[i] represents the weight for xdata[i]. If weight_x is an array of shape (m,), then it represents the diagonal of the covariant weighting matrix for all data points. If weight_x is an array of shape (m, m), then it represents the full covariant weighting matrix for all data points. If weight_x is an array of shape (n, m), then weight_x[i, :] represents the diagonal of the covariant weighting matrix for xdata[i, :]. If weight_x is an array of shape (n, m, m), then weight_x[i, :, :] represents the full covariant weighting matrix for xdata[i, :]. For a comprehensive description of the options, refer to page 26 of the ODRPACK95 guide. By default, weight_x is set to one for all xdata points.

weight_y::Union{Float64,Vector{Float64},Matrix{Float64},Array{Float64,3},Nothing}=nothing: Scalar or array specifying how the errors on ydata are to be weighted. If weight_y is a scalar, then it is used for all data points. If weight_y is an array of shape (n,) and q==1, then weight_y[i] represents the weight for ydata[i]. If weight_y is an array of shape (q), then it represents the diagonal of the covariant weighting matrix for all data points. If weight_y is an array of shape (q, q), then it represents the full covariant weighting matrix for all data points. If weight_y is an array of shape (n, q), then weight_y[i, :] represents the diagonal of the covariant weighting matrix for ydata[i, :]. If weight_y is an array of shape (n, q, q), then weight_y[i, :, :] represents the full covariant weighting matrix for ydata[i, :]. For a comprehensive description of the options, refer to page 25 of the ODRPACK95 guide. By default, weight_y is set to one for all ydata points.

bounds::Tuple{Union{Vector{Float64},Nothing},Union{Vector{Float64},Nothing}}=(nothing, nothing): Tuple of arrays with the same shape as beta0, specifying the lower and upper bounds of the model parameters. The first array contains the lower bounds, and the second contains the upper bounds. By default, the bounds are set to negative and positive infinity, respectively, for all elements of beta.

task::Symbol=:explicitODR: Specifies the regression task to be performed. :explicitODR solves an orthogonal distance regression problem with an explicit model. :implicitODR handles models defined implicitly. :OLS performs ordinary least squares fitting.

fix_beta::Union{Vector{Bool},Nothing}=nothing: Array with the same shape as beta0, specifying which elements of beta are to be held fixed. true means the parameter is fixed; false means it is adjustable. By default, all elements of beta are set to false.

fix_x::Union{VecOrMat{Bool},Nothing}=nothing: Array with the same shape as xdata, specifying which elements of xdata are to be held fixed. Alternatively, it can be a rank-1 array of shape (m,) or (n,), in which case it will be broadcast along the other axis. true means the element is fixed; false means it is adjustable. By default, in orthogonal distance regression mode, all elements of fix_x are set to false. In ordinary least squares mode (task=:OLS), all xdata values are automatically treated as fixed.

jac_beta!::Union{Function,Nothing}=nothing: Jacobian of the function to be fitted with respect to beta, with the signature jac_beta!(x, beta, jb). It must modify the result jb in-place, with shape (n, np, q). By default, the Jacobian is approximated numerically using the finite difference scheme specified by diff_scheme.

jac_x!::Union{Function,Nothing}=nothing: Jacobian of the function to be fitted with respect to x, with the signature jac_x!(x, beta, jx). It must modify the result jx in-place, with shape (n, m, q). By default, the Jacobian is approximated numerically using the finite difference scheme specified by diff_scheme.

delta0::Union{VecOrMat{Float64},Nothing}=nothing: Array with the same shape as xdata, containing the initial guesses of the errors in the explanatory variable. By default, delta0 is set to zero for all elements of xdata.

diff_scheme::Symbol=:forward: Finite difference scheme used to approximate the Jacobian matrices when the user does not provide jac_beta! and jac_x!. The default method is forward (:forward) differences. Central (:central) differences are generally more accurate but require one additional f! evaluation per partial derivative.

report::Symbol=:none: Specifies the level of computation reporting. :none disables all output. :short prints a brief initial and final summary. :long provides a detailed initial and final summary. :iteration outputs information at each iteration step in addition to the detailed summaries. This is useful for debugging or monitoring progress.

maxit::Integer=50: Maximum number of allowed iterations.

ndigit::Union{Integer,Nothing}=nothing: Number of reliable decimal digits in the values computed by the model function f! and its Jacobians jac_beta!, and jac_x!. By default, the value is numerically determined by evaluating f!.

taufac::Union{Float64,Nothing}=nothing: Factor ranging from 0 to 1 to initialize the trust region radius. The default value is 1. Reducing taufac may be appropriate if, at the first iteration, the computed results for the full Gauss-Newton step cause an overflow, or cause beta and/or delta to leave the region of interest.

sstol::Union{Float64,Nothing}=nothing: Factor ranging from 0 to 1 specifying the stopping tolerance for the sum of the squares convergence. The default value is eps()^(1/2), where eps() is the machine precision in Float64.

partol::Union{Float64,Nothing}=nothing: Factor ranging from 0 to 1 specifying the stopping tolerance for parameter convergence (i.e., beta and delta). When the model is explicit, the default value is eps()^(2/3), and when the model is implicit, the default value is eps()^(1/3), where eps() is the machine precision in Float64.

step_beta::Union{Vector{Float64},Nothing}=nothing: Array with the same shape as beta0 containing the relative step sizes used to compute the finite difference derivatives with respect to the model parameters. By default, the step size is set internally based on the value of ndigit and the type of finite differences specified by diff_scheme. For additional details, refer to pages 31 and 78 of the ODRPACK95 guide.

step_delta::Union{VecOrMat{Float64},Nothing}=nothing: Array with the same shape as xdata, containing the relative step sizes used to compute the finite difference derivatives with respect to the errors in the explanatory variable. Alternatively, it can be a rank-1 array of shape (m,) or (n,), in which case it will be broadcast along the other axis. By default, step size is set internally based on the value of ndigit and the type of finite differences specified by diff_scheme. For additional details, refer to pages 31 and 78 of the ODRPACK95 guide.

scale_beta::Union{Vector{Float64},Nothing}=nothing: Array with the same shape as beta0 containing the scale values of the model parameters. Scaling is used to improve the numerical stability of the regression, but does not affect the problem specification. Scaling should not be confused with the weighting matrices weight_x and weight_y. By default, the scale is set internally based on the relative magnitudes of beta. For further details, refer to pages 32 and 84 of the ODRPACK95 guide.

scale_delta::Union{VecOrMat{Float64},Nothing}=nothing: Array with the same shape as xdata, containing the scale values of the errors in the explanatory variable. Alternatively, it can be a rank-1 array of shape (m,) or (n,), in which case it will be broadcast along the other axis. Scaling is used to improve the numerical stability of the regression, but does not affect the problem specification. Scaling should not be confused with the weighting matrices weight_x and weight_y. By default, the scale is set internally based on the relative magnitudes of xdata. For further details, refer to pages 32 and 85 of the ODRPACK95 guide.

rptfile::Union{String,Nothing}=nothing: File name for storing the computation reports, as defined by report. By default, the reports are sent to standard output.

errfile::Union{String,Nothing}=nothing: File name for storing the error reports, as defined by report. By default, the reports are sent to standard output.

Returns

OdrResult: An object containing the results of the regression.

References

[1] Jason W. Zwolak, Paul T. Boggs, and Layne T. Watson. Algorithm 869: ODRPACK95: A weighted orthogonal distance regression code with bound constraints. ACM Trans. Math. Softw. 33, 4 (August 2007), 27-es. https://doi.org/10.1145/1268776.1268782

[2] Jason W. Zwolak, Paul T. Boggs, and Layne T. Watson. User's Reference Guide for ODRPACK95, 2005. https://github.com/HugoMVale/odrpack95/blob/main/original/Doc/guide.pdf

Examples

using Odrpack

function f!(x::Vector{Float64}, beta::Vector{Float64}, y::Vector{Float64})
    y .= beta[1] .* exp.(beta[2] .* x)
    return nothing
end

xdata = [0.982, 1.998, 4.978, 6.01]
ydata = [2.7, 7.4, 148.0, 403.0]

beta0 = [2.0, 0.5]
bounds = ([0.0, 0.0], [10.0, 0.9])

sol = odr_fit(
    f!,
    xdata,
    ydata,
    beta0,
    bounds=bounds,
    # rptfile="test_output.txt",
    # report=:long
)

println("Optimized β    :", sol.beta)
println("Optimized δ    :", sol.delta)
println("Optimized x + δ:", sol.xplusd)
 

Odrpack.Auxiliary — Module

Auxiliary functions for odrpack.

This module provides a Julia interface to the auxiliary routines from the Fortran odrpack library.

Odrpack.Auxiliary.close_file — Method

close_file(lun) -> Int32

Close a file associated with a specified logical unit number.

Arguments

lun::Ref{Int32}: The logical unit number of the file to close.

Returns

Int32: An error code (compiler dependent). A return value of 0 typically indicates successful closure.

Odrpack.Auxiliary.loc_iwork — Method

loc_iwork(m, q, np) -> Iworkidx

Get storage locations within the integer work space.

Arguments

m::Integer: Number of columns of data in the explanatory variable.

q::Integer: Number of responses per observation.

np::Integer: Number of function parameters.

Returns

Iworkidx: A structure containing the 0-based indexes of the integer work array.

Odrpack.Auxiliary.loc_rwork — Method

loc_rwork(n, m, q, np, ldwe, ld2we, isodr) -> Rworkidx

Get storage locations within the real work space.

Arguments

n::Integer: Number of observations.

m::Integer: Number of columns of data in the explanatory variable.

q::Integer: Number of responses per observation.

np::Integer: Number of function parameters.

ldwe::Integer: Leading dimension of the we array.

ld2we::Integer: Second dimension of the we array.

isodr::Bool: Indicates whether the solution is by ODR (true) or by OLS (false).

Returns

Rworkidx: A structure containing the 0-based indexes of the real work array.

Odrpack.Auxiliary.open_file — Method

open_file(filename, lun) -> Int32

Open a new file associated with a specified logical unit number.

Arguments

filename::AbstractString: String containing the file name.

lun::Ref{Cint}: Logical unit number. This value will be modified by the function.

Returns

Int32: Error code (compiler dependent). A value of 0 typically indicates success.

Odrpack.Auxiliary.stop_message — Method

stop_message(info) -> String

Return a human-readable message based on the stopping condition returned by odrpack in the info argument of the result.

Arguments

info::Integer: value of the info argument returned by odrpack. This value is used to determine the stopping condition.

Returns

::String: human-readable string describing the stopping condition.

Odrpack.Auxiliary.workspace_dimensions — Method

workspace_dimensions(n, m, q, np, isodr) -> Tuple{Int32, Int32}

Calculate the dimensions of the workspace arrays required by the underlying Fortran library.

Arguments

n::Integer: Number of observations.

m::Integer: Number of columns of data in the explanatory variable.

q::Integer: Number of responses per observation.

np::Integer: Number of function parameters.

isodr::Bool: Variable designating whether the solution is by ODR (true) or by OLS (false).

Returns

Tuple{Int32, Int32}: A tuple containing the lengths of the real and integer work arrays (lrwork, liwork).